refactored App.tsx into more components
